requirements.txt
crispy-bootstrap4==2024.1
Django==5.0.3
django-crispy-forms==2.1
django-model-utils==4.4.0

